rename service to services
